requirements.txt
# Core Backend
fastapi
uvicorn
python-dotenv
requests

# AI & Data Processing
google-generativeai
pdfplumber

# Optional / Legacy (Keep if needed for future ML)
numpy
pandas

